s28: do not import HA camera at package load
CameraEntityDescription pulled homeassistant.components.camera into
const.py, which CI cannot import (turbojpeg/numpy/av). Descriptions
and picker options live in gear_guard.py; camera.py still loads the
camera platform. Test/load-test install HA's camera+stream pins, not
ours.
